Add cross-platform deserialization tests. Fix spelling of test class name.
